+++ src/SnapOfDisintegration.js
@@ -22,10 +22,13 @@ const Me             = imports.misc.extensionUtils.getCurrentExtension();
 const utils          = Me.imports.src.utils;
 
 //////////////////////////////////////////////////////////////////////////////////////////
-// This effect shatters the window into pieces. For an explanation how this works, look //
-// further down in this file into the documentation of the Shader class. This effect is //
-// not available on GNOME 3.3x, due to the limitation described in the documentation    //
-// of vfunc_paint_target further down in this file.                                     //
+// This effects dissolves your windows into a cloud of dust. For this, it uses an       //
+// approach similar to the Broken Glass effect. A dust texture is used to segment the   //
+// window texture into a set of layers. Each layer is then moved, sheared and scaled    //
+// randomly. Just a few layers are sufficient to create the illusion of many individual //
+// dust particles.                                                                      //
+// This effect is not available on GNOME 3.3x, due to the limitation described in the   //
+// documentation of vfunc_paint_target further down in this file.                       //
 //////////////////////////////////////////////////////////////////////////////////////////
 
 // The shader class for this effect is registered further down in this file.
@@ -34,7 +37,7 @@ let Shader = null;
 // The effect class is completely static. It can be used to get some metadata (like the
 // effect's name or supported GNOME Shell versions), to initialize the respective page of
 // the settings dialog, as well as to create the actual shader for the effect.
-var BrokenGlass = class BrokenGlass {
+var SnapOfDisintegration = class SnapOfDisintegration {
 
   // ---------------------------------------------------------------------------- metadata
 
@@ -48,13 +51,13 @@ var BrokenGlass = class BrokenGlass {
   // effect is enabled currently (e.g. '*-close-effect'), and its animation time
   // (e.g. '*-animation-time').
   static getNick() {
-    return 'broken-glass';
+    return 'snap';
   }
 
   // This will be shown in the sidebar of the preferences dialog as well as in the
   // drop-down menus where the user can choose the effect.
   static getLabel() {
-    return _('Broken Glass');
+    return _('Snap of Disintegration');
   }
 
   // -------------------------------------------------------------------- API for prefs.js
@@ -65,17 +68,15 @@ var BrokenGlass = class BrokenGlass {
   static getPreferences(dialog) {
 
     // Add the settings page to the builder.
-    dialog.getBuilder().add_from_resource('/ui/gtk4/BrokenGlass.ui');
+    dialog.getBuilder().add_from_resource('/ui/gtk4/SnapOfDisintegration.ui');
 
     // Bind all properties.
-    dialog.bindAdjustment('broken-glass-animation-time');
-    dialog.bindAdjustment('broken-glass-scale');
-    dialog.bindAdjustment('broken-glass-gravity');
-    dialog.bindAdjustment('broken-glass-blow-force');
-    dialog.bindSwitch('broken-glass-use-pointer');
+    dialog.bindAdjustment('snap-animation-time');
+    dialog.bindAdjustment('snap-scale');
+    dialog.bindColorButton('snap-color');
 
     // Finally, return the new settings page.
-    return dialog.getBuilder().get_object('broken-glass-prefs');
+    return dialog.getBuilder().get_object('snap-prefs');
   }
 
   // ---------------------------------------------------------------- API for extension.js
@@ -94,13 +95,13 @@ var BrokenGlass = class BrokenGlass {
   // current value of the opacity transition is passed as uProgress to the shader.
   // Tweaking the actor's scale during the transition only works properly for GNOME 3.38+.
 
-  // For this effect, windows are set to twice their original size, so that
-  // we have some space to draw the shards. We also set the animation mode to "Linear".
+  // For this effect, windows are set to 1.2 times their original size, so that we have
+  // some space to draw the dust particles. We also set the animation mode to "Linear".
   static tweakTransition(actor, settings, forOpening) {
     return {
       'opacity': {from: 255, to: 255, mode: 1},
-      'scale-x': {from: 2.0, to: 2.0, mode: 1},
-      'scale-y': {from: 2.0, to: 2.0, mode: 1}
+      'scale-x': {from: 1.2, to: 1.2, mode: 1},
+      'scale-y': {from: 1.2, to: 1.2, mode: 1}
     };
   }
 }
@@ -117,40 +118,19 @@ if (utils.isInShellProcess()) {
   const {Clutter, GdkPixbuf, Cogl} = imports.gi;
   const shaderSnippets             = Me.imports.src.shaderSnippets;
 
-  // This shader creates a complex-looking effect with rather simple means. Here is how it
-  // works: The window is drawn five times on top of each other (see the SHARD_LAYERS
-  // constant in the GLSL code). Each layer only draws some of the shards, all layers
-  // combined make up the entire window. The layers are then scaled, rotated, and moved
-  // independently from each other - this creates the impression that all shards are
-  // moving independently. In reality, there are only five groups of shards! Which shard
-  // belongs to which layer is defined by the green channel of the texture
-  // resources/img/shards.png. The red channel of the texture contains the distance to the
-  // shard edges. This information is used to fade out the shards.
   Shader = GObject.registerClass({}, class Shader extends Clutter.ShaderEffect {
     _init(actor, settings, forOpening) {
       super._init({shader_type: Clutter.ShaderType.FRAGMENT_SHADER});
 
-      // Load the shards texture. As the shader is re-created for each window animation,
+      // Load the dust texture. As the shader is re-created for each window animation,
       // this texture is also re-created each time. This could be improved in the future!
-      const shardData    = GdkPixbuf.Pixbuf.new_from_resource('/img/shards.png');
-      this._shardTexture = new Clutter.Image();
-      this._shardTexture.set_data(shardData.get_pixels(), Cogl.PixelFormat.RGB_888,
-                                  shardData.width, shardData.height, shardData.rowstride);
+      const dustData    = GdkPixbuf.Pixbuf.new_from_resource('/img/dust.png');
+      this._dustTexture = new Clutter.Image();
+      this._dustTexture.set_data(dustData.get_pixels(), Cogl.PixelFormat.RGB_888,
+                                 dustData.width, dustData.height, dustData.rowstride);
 
-      // Usually, the shards fly away from the center of the window.
-      let epicenterX = 0.5;
-      let epicenterY = 0.5;
-
-      // However, if this option is set, we use the mouse pointer position.
-      if (!forOpening && settings.get_boolean('broken-glass-use-pointer')) {
-        const [x, y]               = global.get_pointer();
-        const [ok, localX, localY] = actor.transform_stage_point(x, y);
-
-        if (ok) {
-          epicenterX = localX / actor.width;
-          epicenterY = localY / actor.height;
-        }
-      }
+      // The dust particles will fade to this color over time.
+      const color = Clutter.Color.from_string(settings.get_string('snap-color'))[1];
 
       // If we are currently performing integration test, the animation uses a fixed seed.
       const testMode = settings.get_boolean('test-mode');
@@ -159,89 +139,90 @@ if (utils.isInShellProcess()) {
 
         // Inject some common shader snippets.
         ${shaderSnippets.standardUniforms()}
+        ${shaderSnippets.math2D()}
 
-        uniform sampler2D uShardTexture;
+        uniform sampler2D uDustTexture;
 
         const vec2  SEED         = vec2(${testMode ? 0 : Math.random()}, 
                                         ${testMode ? 0 : Math.random()});
-        const float SHARD_SCALE  = ${settings.get_double('broken-glass-scale')};
-        const float SHARD_LAYERS = 5;
-        const float BLOW_FORCE   = ${settings.get_double('broken-glass-blow-force')};
-        const float ACTOR_SCALE  = 2.0;
+        const float DUST_SCALE   = ${settings.get_double('snap-scale')};
+        const float DUST_LAYERS  = 5;
+        const float ACTOR_SCALE  = 1.2;
         const float PADDING      = ACTOR_SCALE / 2.0 - 0.5;
-        const vec2  EPICENTER    = vec2(${epicenterX}, ${epicenterY});
-        const float GRAVITY      = ${forOpening ? '-1.0' : '1.0'} * 
-                                   ${settings.get_double('broken-glass-gravity')};
-
+        const vec2  EPICENTER    = vec2(0.5);
+        const vec3  DUST_COLOR   = vec3(${color.red / 255},
+                                        ${color.green / 255},
+                                        ${color.blue / 255});
         void main() {
+
+          // We simply inverse the progress for opening windows.
+          float progress = pow(${forOpening ? '1.0-uProgress' : 'uProgress'}, 2.0);
 
           cogl_color_out = vec4(0, 0, 0, 0);
 
-          float progress = ${forOpening ? '1.0-uProgress' : 'uProgress'};
-
-          // Draw the individual shard layers.
-          for (float i=0; i<SHARD_LAYERS; ++i) {
-
-            // To enable drawing shards outside of the window bounds, the actor was scaled
-            // by ACTOR_SCALE. Here we scale and move the texture coordinates so that the
-            // window gets drawn at the correct position again.
-            vec2 coords = cogl_tex_coord_in[0].st * ACTOR_SCALE - PADDING;
-
-            // Scale and rotate around our epicenter.
-            coords -= EPICENTER;
-
-            // Scale each layer a bit differently.
-            coords /= mix(1.0, 1.0 + BLOW_FORCE*(i+2)/SHARD_LAYERS, progress);
+          for (float i=0; i<DUST_LAYERS; ++i) {
             
-            // Rotate each layer a bit differently.
-            float rotation = (mod(i, 2.0)-0.5)*0.2*progress;
-            coords = vec2(coords.x * cos(rotation) - coords.y * sin(rotation),
-            coords.x * sin(rotation) + coords.y * cos(rotation));
+            // Create a random direction.
+            float factor   = DUST_LAYERS == 1 ? 0 : i/(DUST_LAYERS-1);
+            float angle    = 123.123 * (SEED.x + factor);
+            vec2 direction = vec2(1.0, 0.0);
+            direction      = rotate(direction, angle);
             
-            // Move down each layer a bit.
-            float gravity = GRAVITY*0.1*(i+1)*progress*progress;
-            coords += vec2(0, gravity);
+            // Flip direction for one side of the window.
+            vec2 coords = cogl_tex_coord_in[0].st * ACTOR_SCALE - PADDING - EPICENTER;
+            if (getWinding(direction, coords) > 0) {
+              direction *= -1;
+            }
 
-            // Restore correct position.
-            coords += EPICENTER;
+            // Flip direction for half the layers.
+            if (factor > 0.5) {
+              direction *= -1;
+            }
             
-            // Retrieve information from the shard texture for our layer.
-            vec2 shardCoords = (coords + SEED) * vec2(uSizeX, uSizeY) / SHARD_SCALE / 500.0;
-            vec2 shardMap = texture2D(uShardTexture, shardCoords).rg;
+            // We grow the layer along the random direction, shrink it orthogonally to it
+            // and scale it up slightly.
+            float dist  = distToLine(vec2(0.0), direction, coords);
+            vec2 grow   = direction * dist * mix(0, 0.1, progress);
+            vec2 shrink = vec2(direction.y, -direction.x) * dist * mix(0, 0.1, progress);
+            float scale = mix(1.0, 1.1, factor * progress);
 
-            // The green channel contains a random value in [0..1] for each shard. We
-            // discretize this into SHARD_LAYERS bins and check if our layer falls into
-            // the bin of the current shard.
-            float shardGroup = floor(shardMap.g * SHARD_LAYERS * 0.999);
+            coords = (coords + grow + shrink) / scale;
+         
+            // Now check wether there is actually something in the current dust layer at
+            // the coords position.
+            vec2 dustCoords = (coords + SEED) * vec2(uSizeX, uSizeY) / DUST_SCALE / 100.0;
+            vec2 dustMap    = texture2D(uDustTexture, dustCoords).rg;
+            float dustGroup = floor(dustMap.g * DUST_LAYERS * 0.999);
 
-            if (shardGroup == i) {
-              vec4 windowColor = texture2D(uTexture, coords);
+            if (dustGroup == i) {
 
-              // Dissolve the shard by using distance information from the red channel.
-              float dissolve = (shardMap.x - pow(progress+0.1, 2)) > 0 ? 1: 0;
-              cogl_color_out = mix(cogl_color_out, windowColor, dissolve);
+              // Fade the window color to DUST_COLOR.
+              vec4 windowColor = texture2D(uTexture, coords + EPICENTER);
+              windowColor.rgb = mix(windowColor.rgb, DUST_COLOR*windowColor.a, progress);
+
+              // Dissolve the dust particles.
+              float dissolve = (dustMap.x - progress) > 0 ? 1 : 0;
+              windowColor *= dissolve;
+
+              // Blend the layers.
+              cogl_color_out = mix(cogl_color_out, windowColor, windowColor.a);
             }
           }
         }
       `);
     };
 
-    // This is overridden to bind the shard texture for drawing. Sadly, this seems to be
+    // This is overridden to bind the dust texture for drawing. Sadly, this seems to be
     // impossible under GNOME 3.3x as this.get_pipeline() is not available. It was called
     // get_target() back then but this is not wrapped in GJS.
     // https://gitlab.gnome.org/GNOME/mutter/-/blob/gnome-3-36/clutter/clutter/clutter-offscreen-effect.c#L598
     vfunc_paint_target(node, paint_context) {
       const pipeline = this.get_pipeline();
-
-      // Use linear filtering for the window texture.
       pipeline.set_layer_filters(0, Cogl.PipelineFilter.LINEAR,
                                  Cogl.PipelineFilter.LINEAR);
-
-      // Bind the shard texture.
-      pipeline.set_layer_texture(1, this._shardTexture.get_texture());
+      pipeline.set_layer_texture(1, this._dustTexture.get_texture());
       pipeline.set_layer_wrap_mode(1, Cogl.PipelineWrapMode.REPEAT);
-      this.set_uniform_value('uShardTexture', 1);
-
+      this.set_uniform_value('uDustTexture', 1);
       super.vfunc_paint_target(node, paint_context);
     }
   });